+++ test/schema_validator/hyper_schema/response_validator_test.rb
@@ -1,4 +1,4 @@
-require_relative "test_helper"
+require_relative "../../test_helper"
 
 describe Committee::SchemaValidator::HyperSchema::ResponseValidator do
   before do
